routed all controller methods to their common files
