+++ routes/paymentroutes.js
@@ -2,6 +2,7 @@ const router = require("express").Router();
 const { ensureAuth } = require("../middleware/auth");
 const bodyParser = require("body-parser");
 const cookieParser = require("cookie-parser");
+const paymentControllers = require('../controllers/payments');
 
 
 router.use(cookieParser());
@@ -10,19 +11,8 @@ router.use(bodyParser.json());
 
 
 //route for completed payment
-router.get("/complete", ensureAuth, (req, res) => {
-  const user = req.user;
-  if (!req.query.reference) {
-    res.redirect("/");
-  }
-  res.render("payment/payment_complete", {
-    user,
-  });
-});
+router.get("/complete", ensureAuth, paymentControllers.paymentCompletion);
 
-router.post("/complete-transaction", ensureAuth, (req, res) => {
-  console.log("body", req.body);
-  return res.status(200).json({ data: req.body });
-});
+router.post("/complete-transaction", ensureAuth, paymentControllers.paymentWebhook);
 
 module.exports = router;
